+++ src/main/java/co/tryterra/terraclient/models/v2/samples/BreathSample.java
@@ -30,5 +30,5 @@ import lombok.Setter;
 public class BreathSample {
     private String timestamp;
     @JsonProperty("breaths_per_min")
-    private float breathsPerMin;
+    private double breathsPerMin;
 }
